+++ src/Relations/BelongsToMany.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace Codemonster\Database\Relations;
+
+use Codemonster\Database\ORM\Model;
+use Codemonster\Database\ORM\ModelCollection;
+
+class BelongsToMany extends Relation
+{
+    protected string $pivotTable;
+    protected string $foreignPivotKey;
+    protected string $relatedPivotKey;
+    protected string $parentKey;
+    protected string $relatedKey;
+
+    public function __construct(
+        $builder,
+        Model $parent,
+        Model $related,
+        string $pivotTable,
+        string $foreignPivotKey,
+        string $relatedPivotKey,
+        string $parentKey,
+        string $relatedKey
+    ) {
+        parent::__construct($builder, $parent, $related);
+
+        $this->pivotTable     = $pivotTable;
+        $this->foreignPivotKey = $foreignPivotKey;
+        $this->relatedPivotKey = $relatedPivotKey;
+        $this->parentKey      = $parentKey;
+        $this->relatedKey     = $relatedKey;
+    }
+
+    public function getResults(): ModelCollection
+    {
+        $parentId    = $this->parent->{$this->parentKey};
+        $relatedTable = $this->related->getTable();
+
+        $rows = $this->builder
+            ->join(
+                $this->pivotTable,
+                $this->pivotTable . '.' . $this->relatedPivotKey,
+                '=',
+                $relatedTable . '.' . $this->relatedKey
+            )
+            ->where($this->pivotTable . '.' . $this->foreignPivotKey, $parentId)
+            ->select($relatedTable . '.*')
+            ->get();
+
+        $class = get_class($this->related);
+
+        /** @var class-string<Model> $class */
+        return $class::hydrate($rows);
+    }
+}