+++ src/Relations/Relation.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Codemonster\Database\Relations;
+
+use Codemonster\Database\ORM\Model;
+use Codemonster\Database\ORM\ModelCollection;
+use Codemonster\Database\Query\QueryBuilder;
+
+abstract class Relation
+{
+    protected QueryBuilder $builder;
+    protected Model $parent;
+    protected Model $related;
+
+    public function __construct(QueryBuilder $builder, Model $parent, Model $related)
+    {
+        $this->builder = $builder;
+        $this->parent  = $parent;
+        $this->related = $related;
+    }
+
+    abstract public function getResults();
+
+    protected function hydrate(array $rows): ModelCollection
+    {
+        $class = get_class($this->related);
+
+        /** @var class-string<Model> $class */
+        return $class::hydrate($rows);
+    }
+}